Move public/debugger/assets ts files
